Minor changes from coverity code analysis Added Docs/Solution Layout.md

+++ Src/Libs/Util/Graphics/Image.cpp
@@ -321,8 +321,8 @@ namespace Graphics
 
     uint64_t Image::calcTrimSavings(const Rect& rect) const
     {
-        uint64_t savings = rect.width < _size.width ? (_size.width - rect.width) * rect.height : 0;
-        savings += rect.height < _size.height ? (_size.height - rect.height) * _size.width : 0;
+        uint64_t savings = rect.width < _size.width ? static_cast<uint64_t>(_size.width - rect.width) * rect.height : 0;
+        savings += rect.height < _size.height ? static_cast<uint64_t>(_size.height - rect.height) * _size.width : 0;
         return savings * sizeof(Pixel);
     }
